% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_sb.R
\name{sa_sb}
\alias{sa_sb}
\title{Titration of Strong Acid with Strong Base.}
\usage{
sa_sb(ca = 0.1, cb = 0.1, pkw = 14, va = 50, eqpt = F, overlay = F)
}
\arguments{
\item{ca}{Molar concentration of strong acid analyte; defaults to
0.10 M.}

\item{cb}{Molar concentration of strong base titrant; defaults to
0.10 M.}

\item{pkw}{pKw value for the solvent; defaults to a pKw of 14.}

\item{va}{Volume in mL of solution containing the analyte; defaults
to 50 mL.}

\item{eqpt}{Show equivalence point (T/F); defaults to FALSE.}

\item{overlay}{Overlay new titration curve onto previous plot
(T/F); defaults to FALSE.}
}
\description{
This function plots the full titration curve for a strong acid
using a strong base as a titrant using the approach outlined in R.
deLevie's \emph{Principles of Quantitative Analysis}. Options
include overlaying two or more titration curves and showing the
location of the equivalence point.
}
\examples{
sa_sb(ca = 0.10, eqpt = T)
sa_sb(ca = 0.05, overlay = T, eqpt = T)
sa_sb(ca = 0.10)
}

