% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbmix_sa.R
\name{wbmix_sa}
\alias{wbmix_sa}
\title{Titration Curve for Weak Base Mixture}
\usage{
wbmix_sa(conc.base1 = 0.1, conc.base2 = 0.1, conc.acid = 0.1, pka1 = 6,
  pka2 = 9, pkw = 14, vol.base = 50, plot = TRUE, eqpt = FALSE,
  overlay = FALSE, ...)
}
\arguments{
\item{conc.base1}{Molar concentration of the first monoprotic 
weak base analyte; defaults to 0.10 M.}

\item{conc.base2}{Molar concentration of the second monoprotic 
weak base analyte; defaults to 0.10 M.}

\item{conc.acid}{Molar concentration of the strong acid titrant;
defaults to 0.10 M.}

\item{pka1}{The pKa value for the first monoprotic weak base 
analyte's conjugate weak acid; defaults to a pKa of 6, or a pKb of
8.}

\item{pka2}{The pKa value for the second monoprotic weak base 
analyte's conjugate weak acid; defaults to a pKa of 9, or a pKb of
5.}

\item{pkw}{The pKw (or pKs) value for the solvent; defaults to water
as a solvent with a pKw of 14.}

\item{vol.base}{Initial volume, in mL, of the solution that 
contains the weak base; defaults to 50.00 mL.}

\item{plot}{Logical; if TRUE, plots the titration curve.}

\item{eqpt}{Logical; if TRUE, draws a vertical line at the titration
curve's equivalence point.}

\item{overlay}{Logical; if TRUE, adds the current titration curve
to the existing titration curve.}

\item{\dots}{Additional arguments to pass to \code{plot()} function.}
}
\value{
A two-column data frame that contains the volume of titrant
in the first column and the solution's pH in the second column. Also
produces a plot of the titration curve with options to display the
equivalence point and to overlay titration curves.
}
\description{
This function calculates and plots the titration curve for a 
mixture of two monoprotic weak base using a monoprotic 
strong acid as the titrant. The calculation uses a single master 
equation that finds the volume of titrant needed to achieve a fixed 
pH, as outlined in R. de Levie's \emph{Principles of Quantitative 
Chemical Analysis} (McGraw-Hill, 1997).
}
\examples{
### Simple titration curve with equivalence points
ex10 = wbmix_sa(eqpt = TRUE)
head(ex10)

### Overlay titration curves using different pKa values
wbmix_sa(pka1 = 5, pka2 = 8, eqpt = TRUE)
wbmix_sa(pka1 = 4, pka2 = 7, overlay = TRUE)
wbmix_sa(pka1 = 6, pka2 = 9, overlay = TRUE)
}
\author{
David T. Harvey, DePauw University. \email{harvey@depauw.edu}
}

